Refactor transformer into its own file
